\name{cran-comments}
\alias{cran-comments}
\title{
Note on notes
}
\description{
This is an updated and extended version of `spm` package. The change in package name from `spm` to `spm2` is due to the change in Author's support from Geoscience Australia to Data2Action Australia.

## R CMD check results
0 errors | 0 warnings | 0 notes
}
\author{
Jin Li
}
